requirements.txt
# requirements.txt - Dependencias del proyecto

# Framework web
Flask==2.3.3
gunicorn==21.2.0

# Cliente OpenAI
openai==1.3.8

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0

# Web scraping (para extraer contenido web)
beautifulsoup4==4.12.2
lxml==4.9.3

# Logging y utilidades
python-dateutil==2.8.2

# Para producción
Werkzeug==2.3.7